+++ Projet/miniprojet/Miniprojet/process_image.h
@@ -1,7 +1,7 @@
 #ifndef PROCESS_IMAGE_H
 #define PROCESS_IMAGE_H
 
-float get_position_px(void);
+uint16_t get_position_px(void);
 void process_image_start(void);
 
 void wait_position_acquired(void);
